@@ -15,7 +15,7 @@
     "dayjs": "^1.11.13",
     "microcms-js-sdk": "^3.1.2",
     "microcms-rich-editor-handler": "^1.0.1",
-    "next": "15.5.9",
+    "next": "^16.1.4",
     "next-auth": "^4.24.11",
     "react": "^19.2.3",
     "react-dom": "^19.2.3",
@@ -35,7 +35,7 @@
     "@vitest/browser": "^3.2.4",
     "@vitest/coverage-v8": "3.2.4",
     "eslint": "^9",
-    "eslint-config-next": "15.5.9",
+    "eslint-config-next": "^16.1.4",
     "eslint-plugin-storybook": "^9.1.5",
     "playwright": "^1.55.0",
     "postcss": "^8",
