@@ -4,7 +4,7 @@ import { BlogsByCategory } from "@/libs/schema/contents/Category/category";
 import BlogCardGrid from "@/commons/blog/components/BlogCardGrid/BlogCardGrid";
 
 type Props = {
-  queryId: string | null;
+  queryId?: string;
   blogCategoryList: BlogsByCategory[];
 };
 
